Refactor handler
