+++ src/backend.ts
@@ -2,7 +2,6 @@ export * from './backend/Database.js'
 export * from './backend/FS.js'
 export * from './backend/Handler.js'
 export * from './backend/Media.js'
-export * from './backend/Server.js'
 export * from './backend/Target.js'
 export * from './backend/loader/JsonLoader.js'
 export * from './backend/util/JWTPreviews.js'
